feature: Group all service commands under service
