--- cmd/internal/client/bind-service.go
@@ -1,66 +1,0 @@
-package client
-
-import (
-	"github.com/pkg/errors"
-	"github.com/spf13/cobra"
-	"github.com/suse/carrier/paas"
-)
-
-var ()
-
-// CmdBindService implements the carrier bind-service command
-var CmdBindService = &cobra.Command{
-	Use:   "bind-service NAME APP",
-	Short: "Bind a service to an application",
-	Long:  `Bind service by name, to named application.`,
-	Args:  cobra.ExactArgs(2),
-	RunE: func(cmd *cobra.Command, args []string) error {
-		client, cleanup, err := paas.NewCarrierClient(cmd.Flags())
-		defer func() {
-			if cleanup != nil {
-				cleanup()
-			}
-		}()
-
-		if err != nil {
-			return errors.Wrap(err, "error initializing cli")
-		}
-
-		err = client.BindService(args[0], args[1])
-		if err != nil {
-			return errors.Wrap(err, "error binding service")
-		}
-
-		return nil
-	},
-	SilenceErrors: true,
-	SilenceUsage:  true,
-	ValidArgsFunction: func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-		if len(args) > 1 {
-			return nil, cobra.ShellCompDirectiveNoFileComp
-		}
-
-		app, cleanup, err := paas.NewCarrierClient(cmd.Flags())
-		defer func() {
-			if cleanup != nil {
-				cleanup()
-			}
-		}()
-
-		if err != nil {
-			return nil, cobra.ShellCompDirectiveNoFileComp
-		}
-
-		if len(args) == 1 {
-			// #args == 1: app name.
-			matches := app.AppsMatching(toComplete)
-			return matches, cobra.ShellCompDirectiveNoFileComp
-		}
-
-		// #args == 0: service name.
-
-		matches := app.ServiceMatching(toComplete)
-
-		return matches, cobra.ShellCompDirectiveNoFileComp
-	},
-}
