+++ cmd/root.go
@@ -63,15 +63,7 @@ func Execute() {
 	rootCmd.AddCommand(client.CmdTarget)
 	rootCmd.AddCommand(client.CmdEnable)
 	rootCmd.AddCommand(client.CmdDisable)
-	rootCmd.AddCommand(client.CmdCreateCustomService)
-	rootCmd.AddCommand(client.CmdCreateService)
-	rootCmd.AddCommand(client.CmdDeleteService)
-	rootCmd.AddCommand(client.CmdBindService)
-	rootCmd.AddCommand(client.CmdUnbindService)
-	rootCmd.AddCommand(client.CmdServices)
 	rootCmd.AddCommand(client.CmdService)
-	rootCmd.AddCommand(client.CmdServiceClasses)
-	rootCmd.AddCommand(client.CmdServicePlans)
 
 	if err := rootCmd.Execute(); err != nil {
 		fmt.Println(err)
